% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{add_bars_for_pcts}
\alias{add_bars_for_pcts}
\title{Add bars depicting the percents within a summary table}
\usage{
add_bars_for_pcts(x)
}
\arguments{
\item{x}{A gtsummary or gt table data object}
}
\value{
A gt table data object
}
\description{
Add bars depicting the percents within a summary table
}
\examples{
x <- data.frame(
sex = c(rep(1, 75), rep(2, 20), rep(NA_integer_, 5)) |>
  labelled::labelled(
    labels = c(Male = 1, Female = 2, Transwoman = 3),
    label = 'Sex Assigned at Birth'
  )
)
x |>
  Misc.SHH.f::labelled_to_factor() |>
  gtsummary::tbl_summary() |>
  add_bars_for_pcts()
}
