% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_no_rows}
\alias{assert_no_rows}
\title{Assert the absence of rows in a data.frame}
\usage{
assert_no_rows(df, msg = NULL, sound = 9)
}
\arguments{
\item{df}{A data.frame (or, probably more usefully, an expression subsetting a data.frame).}

\item{msg}{An optional message to display if \code{df} has any rows.}

\item{sound}{A sound to play if the assertion fails. See \code{\link[beepr:beep]{beepr::beep()}} for a
list of valid options.}
}
\value{
If the assertion fails, generates an error message. If run
interactively, it will also play the specified sound to call the user's
attention to the fact that the calling R session has encountered an error
and halted code execution, display the error in a dialog, and open a new
data viewer tab displaying the offending rows.
}
\description{
Asserts the absence of rows in a data.frame. Particularly useful for asserting
the absence of rows \strong{meeting some specified criteria} by passing \strong{an
expression that subsets a data.frame} (see examples).
}
\examples{
x <- data.frame(x = c(rep(1, 5), 6:10), y = c(rep(11, 3), 14:20))

assert_no_rows(x |> dplyr::filter(y == 14))

# throw an error with a custom message
assert_no_rows(
  x |> dplyr::filter(y == 14),
  msg = "y is 14 in some rows of x"
)
}
