% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_vars_in_df}
\alias{assert_vars_in_df}
\title{Check that 1+ variables are present in a data.frame}
\usage{
assert_vars_in_df(df, vars)
}
\arguments{
\item{vars}{A character vector containing 1+ names of columns expected in df}
}
\value{
An error if any variable specified in \code{vars} is not present in \code{df}
}
\description{
#' @param df A data frame
}
\examples{
data.frame(x = letters[1:3], y = 1:3) |> assert_vars_in_df(c("y", "u", "v"))
data.frame(x = letters[1:3], y = 1:3, u = 4:6) |> assert_vars_in_df(c("y", "u", "v"))
data.frame(x = letters[1:3], y = 1:3, u = 4:6, v = 7:9) |> assert_vars_in_df(c("y", "u", "v"))
}
