% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_var_note.R
\name{gen_data_dict}
\alias{gen_data_dict}
\title{Generate a data dictionary for a data frame}
\usage{
gen_data_dict(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A data frame with columns Variable, Label, Type, Classes, Values, and
Notes describing each column in \code{df}
}
\description{
Generate a data dictionary for a data frame
}
\examples{
x <- data.frame(
  x = rep(1:3,2) |>
    labelled::labelled(labels = c("One" = 1, "Two" = 2, "Three" = 3)) |>
    labelled::set_variable_labels("An integer") |>
    set_notes_attr("These are just integers between 1 and 3"),
  y = rep(letters[1:3],2) |>
    labelled::labelled(labels = c("A" = "a", "B" = "b", "C" = "c")) |>
    labelled::set_variable_labels("A letter") |>
    set_notes_attr("These are just letters 'a' through 'c'")
  ) |>
  set_notes_attr("An example dataset")

gen_data_dict(x) |> write.csv("temp.csv")
}
