% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{group_dichot_vars}
\alias{group_dichot_vars}
\title{Group breakdowns by non-mutually exclusive race/ethnicity variables (e.g.,
natam_multi, black_multi, latino)}
\usage{
group_dichot_vars(
  x,
  variables,
  header,
  pattern_label_to_omit = "^((NOT)|(NO))\\\\b.+",
  pattern_label_to_show = NULL
)
}
\arguments{
\item{x}{gtsummary object of class 'tbl_summary' containing breakdowns by the
variables specified in \code{vars} (possibly among other variables)}

\item{variables}{A character vector of names of variables to group that appear
in x$table_body. Selected variables should be appear consecutively in table}

\item{header}{A string specifying the header/label to place above the variable
group (e.g., "Other factors")}

\item{pattern_label_to_omit}{String specifying a regular expression matching
labels (converted to uppercase for purposes of comparison) for categories to
omit from the variable group. Defaults to "^((NOT)|(NO))\\b.+", matching
labels beginning with "No" or "Not" (e.g., "Not Black", "No documented meth
use")}

\item{pattern_label_to_show}{String specifying a regular expression matching
labels (converted to uppercase for purposes of comparison) for categories to
show in the variable group. Defaults to NULL. If specified,
\code{pattern_label_to_omit} is ignored--set it to NULL to suppress a warning
about this.}
}
\value{
A gtsummary table object. If any of the tabulated observations (e.g.,
PWH) have an unknown/missing value for a variable included in the variable
group, a footnote is added to the label for that variable specifying the
number missing
}
\description{
Group breakdowns by non-mutually exclusive race/ethnicity variables (e.g.,
natam_multi, black_multi, latino)
}
\examples{
df <- kcHARS::read_eoq_append(
date = "2026-07-01",
col_select = c(ethnicity1, matches("race[0-9]"), birth_country_cd)
) |>
  kcHARS::calc_race_eth_vars() |>
  kcHARS::calc_nativity() |>
  dplyr::select(-ethnicity1, -matches("race[0-9]"), -birth_country_cd)

x <- df |> Misc.SHH.f::labelled_to_factor() |> gtsummary::tbl_summary()

x # what the table looks like BEFORE applying group_race_eth_multi_vars()

x |>
  group_dichot_vars(
    variables = c(
      "black_multi",
      "natam_multi",
      "asian_multi",
      "black_multi",
      "latino",
      "pacisl_multi",
      "white_multi",
      "nativity"
    ),
    header = "Misc",
    pattern_label_to_omit = "^(((NOT)|(NO))\\\\b.)|(U\\\\.S\\\\.\\\\-BORN)+"
  )
}
