% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_to_factor.R
\name{labelled_to_factor}
\alias{labelled_to_factor}
\title{Convert all labelled columns to factors (e.g., in preparation for generating
tables and plots)}
\usage{
labelled_to_factor(df, droplevels = TRUE)
}
\arguments{
\item{df}{A data frame}

\item{droplevels}{Whether to drop unused factor levels, such that so that
categories unrepresented in the data (i.e., with 0 counts) are omitted from
summary tables. Defaults to TRUE.}

\item{ordered}{Whether or not the factors created should be ordered factors.
Defaults to TRUE. If so, the order of factor levels matches the order in
which the associated labels were defined.}
}
\value{
Input data frame \code{df} with labelled columns converted to factors.
When tabulating or plotting data, results in categories appearing in the
order in which the associated labels were defined.
}
\description{
Convert all labelled columns to factors (e.g., in preparation for generating
tables and plots)
}
\examples{
x <- data.frame(sex = c(1,1,2,3) |> labelled::labelled(labels = c(Male = 1, Female = 2, Transwoman = 3), label = 'Sex Assigned at Birth'))
str(x)
str(labelled_to_factor(x))
}
