% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{refmt_pcts_lt1}
\alias{refmt_pcts_lt1}
\title{Reformat percents less 1 so they show as "<1\%"}
\usage{
refmt_pcts_lt1(x)
}
\arguments{
\item{x}{A gtsummary data object}
}
\value{
A gt table data object
}
\description{
Reformat percents less 1 so they show as "<1\%"
}
\examples{
x <- data.frame(
sex = c(rep(1, 750), rep(2, 200), rep(NA_integer_, 42), rep(3, 8)) |>
  labelled::labelled(
    labels = c(Male = 1, Female = 2, Transwoman = 3),
  )
)
# before
x |>
  Misc.SHH.f::labelled_to_factor() |>
  gtsummary::tbl_summary() |>
  Misc.SHH.f::add_bars_for_pcts()
# after
x |>
  Misc.SHH.f::labelled_to_factor() |>
  gtsummary::tbl_summary() |>
  refmt_pcts_lt1() |>
  Misc.SHH.f::add_bars_for_pcts()
}
