% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_var_note.R
\name{set_notes_attr}
\alias{set_notes_attr}
\title{Attach \code{notes} attribute}
\usage{
set_notes_attr(x, note)
}
\arguments{
\item{x}{Any arbitrary object (e.g., a vector/variable, a data frame)}

\item{note}{A note to associate with that object}
}
\value{
\code{x} (+ an additional/modified \code{notes} attribute)
}
\description{
Attach \code{notes} attribute
}
\examples{
x <- data.frame(
  x = rep(1:3,2) |>
    labelled::labelled(labels = c("One" = 1, "Two" = 2, "Three" = 3)) |>
    labelled::set_variable_labels("An integer") |>
    set_notes_attr("These are just integers between 1 and 3"),
  y = rep(letters[1:3],2) |>
    labelled::labelled(labels = c("A" = "a", "B" = "b", "C" = "c")) |>
    labelled::set_variable_labels("A letter") |>
    set_notes_attr("These are just letters 'a' through 'c'")
  ) |>
  set_notes_attr("An example dataset")
}
