% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_unique}
\alias{assert_unique}
\title{Assert the uniqueness of some variable(s) in a data frame}
\usage{
assert_unique(df, by = NULL, sound = 9)
}
\arguments{
\item{df}{A data.frame or expression evaluating to one.}

\item{by}{A character vector specifying the name of a column whose values are expected to be unique within \code{df} (or names of 2+ columns, combinations of whose values are expected to be unique).}
}
\value{
If the assertion fails, generates an error message. If run
interactively, will also play the specified sound to call the user's
attention to the fact that the calling R session has encountered an error
and halted code execution, display the error in a dialog, and open a new
data viewer tab displaying the offending rows.
}
\description{
Assert the uniqueness of some variable(s) in a data frame
}
\examples{
x <- data.frame(x = c(rep(1, 5), 6:10), y = c(rep(11, 3), 14:20))
assert_unique(x, by = "x")
assert_unique(x, by = c("x", "y"))
}
